Green do_not_delete_offer_and_supplement_configuration_pax_when_code_is_empty

+++ src/Domain/UseCases/OfferAndSupplementConfigurationPax/DeleteOfferAndSupplementConfigurationPax.cs
@@ -8,6 +8,9 @@ public class DeleteOfferAndSupplementConfigurationPax{
     }
 
     public async Task Execute(string code) {
+        if (code == "") {
+            throw new ArgumentException("Code is required");
+        }
         await availabilitySynchronizerApiClient.DeleteOfferAndSupplementConfigurationPax(code);
     }
 }
